+++ 3rdparty/Jint.CommonJS/Engine.Extensions.cs
@@ -1,7 +1,7 @@
 namespace Jint.CommonJS {
     public static class EngineExtensions {
-        public static ModuleLoadingEngine CommonJS(this Jint.Engine e) {
-            return new ModuleLoadingEngine(e);
+        public static ModuleLoadingEngine CommonJS(this Jint.Engine e, string workingDir) {
+            return new ModuleLoadingEngine(e, workingDir);
         }
     }
 }